+++ src/app/item/item.component.ts
@@ -256,8 +256,8 @@ export class ItemComponent implements OnInit, AfterViewInit, OnChanges, OnDestro
     this.itemList = this.itemList || [];
     if(this.itemList.length === 0) {
       this.itemList.push({text: 'Item 0', type: 'string'});
-      this.itemData = this.itemList[0];
     }
+    this.itemData = this.itemList[0];
     if(this.treeComponent?.treeModel) {
       this.treeComponent?.treeModel.update();
     }
@@ -267,7 +267,15 @@ export class ItemComponent implements OnInit, AfterViewInit, OnChanges, OnDestro
    * Inform the change to host element.
    */
   itemChanged(item) {
-    this.itemData = this.itemData ? Object.assign(this.itemData, item) : null;
+    if(this.itemData) {
+      for (const key of Object.keys(this.itemData)) {
+        if(key !== 'item') {
+          delete this.itemData[key];
+        }
+      }
+      Object.assign(this.itemData, item);
+    }
+
     if (typeof this.itemData?.linkId === 'number') {
       this.itemData.linkId = ''+this.itemData.linkId;
     }