+++ src/app/lib/util.ts
@@ -270,7 +270,10 @@ export class Util {
   static convertToQuestionnaireJSON(value) {
     traverse(value).forEach(function (node) {
       this.before(function () {
-        if (node?.__$helpText?.trim().length > 0) {
+        if(node && Array.isArray(node) && node.includes(undefined)) {
+          this.update(node.filter(()=>{return true}));
+        }
+        else if (node?.__$helpText?.trim().length > 0) {
           const index = Util.findItemIndexWithHelpText(node.item);
           let helpTextItem;
           if (index < 0) {
@@ -432,7 +435,7 @@ export class Util {
    * @param limit - Length to limit the truncation.
    */
   static truncateString(text: string, limit: number = 15): string {
-    return text.length > limit ? (text.substring(0, limit).trim() + '...') : text;
+    return text?.length > limit ? (text.substring(0, limit).trim() + '...') : text;
   }