+++ src/app/lib/widgets/boolean-radio/boolean-radio.component.ts
@@ -11,7 +11,10 @@ import {LfbControlWidgetComponent} from '../lfb-control-widget/lfb-control-widge
   `]
 })
 export class BooleanRadioComponent  extends LfbControlWidgetComponent implements AfterViewInit {
+  static ID = 0;
+  _id = BooleanRadioComponent.ID++;
   ngAfterViewInit() {
+    console.log(this.formProperty.canonicalPathNotation, '=', this.formProperty.value);
     this.labelPosition = 'left';
     super.ngAfterViewInit();
   }